[PCX-1930] Create BrowserController, extract 3DS browser window there
