+++ Sources/UI/Scenes/Input/Protocols/ModalPresenter.swift
@@ -1,0 +1,11 @@
+// Copyright (c) 2021 Payoneer Germany GmbH
+// https://www.payoneer.com
+//
+// This file is open source and available under the MIT license.
+// See the LICENSE file for more information.
+
+import UIKit
+
+protocol ModalPresenter: AnyObject {
+    func present(_ viewControllerToPresent: UIViewController, animated flag: Bool, completion: (() -> Void)?)
+}
